RAG_Pipeline/requirements.txt
langchain-classic
langchain
langchain-community
langchain-core
langchain-mistralai
pypdf
tiktoken
aiofiles
fastapi
uvicorn
jinja2
python-dotenv
python-multipart
PyPDF2
faiss-cpu
pydantic